Tidied up Fragment/IFragment.
