+++ Octokit.GraphQL.Core/Fragment.cs
@@ -4,63 +4,40 @@ using System.Linq.Expressions;
 namespace Octokit.GraphQL
 {
     /// <summary>
-    /// Fragment Interface
+    /// A GraphQL fragment
     /// </summary>
-    public interface IFragment
+    /// <typeparam name="TValue">Input type for the fragment</typeparam>
+    /// <typeparam name="TResult">The output type of the fragment</typeparam>
+    /// <remarks>
+    /// Fragments allow you to share field selections between multiple queries and within the same
+    /// query: https://graphql.org/learn/queries/#fragments
+    /// </remarks>
+    public class Fragment<TValue, TResult> : IFragment
     {
         /// <summary>
-        /// The name of a fragment
+        /// Initializes a new instance of the <see cref="Fragment{TInput, TOutput}"/> class.
         /// </summary>
-        string Name { get; }
-
-        /// <summary>
-        /// The expression to retrieve the value from the result
-        /// </summary>
-        Expression Expression { get; }
-
-        /// <summary>
-        /// Input type for the fragment
-        /// </summary>
-        Type InputType { get; }
-
-        /// <summary>
-        /// The output type of the fragment
-        /// </summary>
-        Type ReturnType { get; }
-    }
-
-    /// <summary>
-    /// Fragment
-    /// </summary>
-    /// <typeparam name="TInput">Input type for the fragment</typeparam>
-    /// <typeparam name="TOutput">The output type of the fragment</typeparam>
-    public class Fragment<TInput, TOutput>: IFragment
-    {
-        /// <inheritdoc />
-        public string Name { get; }
-
-        /// <summary>
-        /// The expression to retrieve the value from the result
-        /// </summary>
-        public Expression<Func<TInput, TOutput>> Expression { get; }
-
-        /// <inheritdoc />
-        Type IFragment.InputType => typeof(TInput);
-        /// <inheritdoc />
-        Type IFragment.ReturnType => typeof(TOutput);
-
-        /// <inheritdoc />
-        Expression IFragment.Expression => Expression;
-
-        /// <summary>
-        /// Create a Fragment
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="expression"></param>
-        public Fragment(string name, Expression<Func<TInput, TOutput>> expression)
+        /// <param name="name">The name of the fragment.</param>
+        /// <param name="expression">The fragment selector expression.</param>
+        public Fragment(string name, Expression<Func<TValue, TResult>> expression)
         {
             Name = name;
             Expression = expression;
         }
+
+        /// <inheritdoc />
+        public string Name { get; }
+
+        /// <inheritdoc />
+        public Expression<Func<TValue, TResult>> Expression { get; }
+
+        /// <inheritdoc />
+        Type IFragment.InputType => typeof(TValue);
+
+        /// <inheritdoc />
+        Type IFragment.ReturnType => typeof(TResult);
+
+        /// <inheritdoc />
+        Expression IFragment.Expression => Expression;
     }
 }
